+++ src/v1/Errors/UnsupportedStringLengthRange/index.ts
@@ -32,10 +32,5 @@
 // POSSIBILITY OF SUCH DAMAGE.
 //
 
-export * from "./isString";
-export * from "./mustBeString";
-export * from "./validateString";
-export * from "./validateStringLengthBetween";
-export * from "./validateStringMatches";
-export * from "./validateStringStartsWith";
-export * from "./validateStringValue";
+export * from "./UnsupportedStringLengthRangeData";
+export * from "./UnsupportedStringLengthRangeError";