+++ CHANGELOG.md
@@ -21,6 +21,11 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### New
+
+* Added `validateStringLengthBetween()`
+* Added `UnsupportedStringLengthRange` Error
+
 ## v0.3.9
 
 Released Monday, 9th May 2022.